requirements.txt

Django==4.0.6
django-ckeditor==6.1.0
django-cors-headers==3.8.0
django-summernote==0.8.20.0
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2==2.9.1